feat: add module module 🫠
